+++ app/assets/src/components/common/MetadataInput.jsx
@@ -1,36 +1,14 @@
 import React from "react";
 import PropTypes from "prop-types";
-import { compact, get, isArray } from "lodash/fp";
+import { isArray } from "lodash/fp";
 
 import Input from "~/components/ui/controls/Input";
 import Dropdown from "~/components/ui/controls/dropdowns/Dropdown";
-import GeoSearchInputBox from "../ui/controls/GeoSearchInputBox";
+import GeoSearchInputBox, {
+  processLocationSelection,
+} from "~/components/ui/controls/GeoSearchInputBox";
 
 import cs from "./metadata_input.scss";
-
-export const LOCATION_PRIVACY_WARNING =
-  "Changed to county/district level for personal privacy.";
-export const LOCATION_UNRESOLVED_WARNING =
-  "Unresolved plain text location, not shown on maps.";
-
-// For human samples, drop the city part of the name and show a warning.
-// Note that the backend will redo the geosearch for confirmation, so don't
-// modify geo_level here.
-// TODO(jsheu): Consider moving the warnings to the backend and generalizing.
-export const processLocationSelection = (result, isHuman) => {
-  let warning = "";
-  if (isHuman && get("geo_level", result) === "city") {
-    result.name = compact([
-      result.subdivision_name,
-      result.state_name,
-      result.country_name,
-    ]).join(", ");
-    warning = LOCATION_PRIVACY_WARNING;
-  } else if (!result.geo_level) {
-    warning = LOCATION_UNRESOLVED_WARNING;
-  }
-  return { result, warning };
-};
 
 class MetadataInput extends React.Component {
   constructor(props) {
